pyproject.toml
[project]
name = "agentic-rag-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "flashrank>=0.2.10",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "pinecone[asyncio,grpc]>=8.1.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "tesseract>=0.1.3",
    "uvicorn[standard]>=0.42.0",
]
